requirements.txt
streamlit
pandas
python-dotenv
pandasai==0.5.9
openai